Improved mouse wheel motion

+++ FalloutStrategy/Motor2D/j1Input.h
@@ -67,13 +67,13 @@ public:
 	// Get mouse / axis position
 	void GetMousePosition(int &x, int &y);
 	void GetMouseMotion(int& x, int& y);
-	void GetMouseWheel(int& y);
+	void GetMouseWheel(float& y);
 
 private:
 	bool		windowEvents[WE_COUNT];
 	j1KeyState*	keyboard;
 	j1KeyState	mouse_buttons[NUM_MOUSE_BUTTONS];
-	int			mouse_wheel_motion_y;
+	float		mouse_wheel_motion_y;
 	int			mouse_motion_x;
 	int			mouse_motion_y;
 	int			mouse_x;
